UBI: skip pages paired with the EC and VID headers
Use the pairing scheme information to skip pages paired with the EC and
VID headers in order to avoid UBI metadata corruption when the a write
instantiated by the UBI user is interrupted by a power-cut.

This is probably not the best solution, because we are loosing one or
several pages/wunits because of that, but for now it's the easiest way to
keep UBI metadata safe without introducing too much complexity.

This will likely be reworked before the final submission.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>