UBI: split the PEB and LEB information at attach time
The UBI code is currently assuming that a PEB contains no more than one
LEB. This assumption is soon to be invalidated by the introduction
of the LEB consolidation concept (which is needed to reliably support
MLC and TLC NANDs).

Split the PEB and LEB information in two different structures
(ubi_ainf_peb and ubi_ainf_leb) to allow multi-LEB PEBs.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>