UBI: rework the on-flash format versioning logic
Currently, the UBI implementation only supports a single on-flash format
version (version 1).

Rework the version checking/initialization logic to allow the same
implementation to support several on-flash version.

If an empty device is detected, the core picks the latest version
supported by the UBI implementation.

Note that we have moved the version validation code in its own function
so that it can be re-used by both VID and EC header validation.

We also dropped the ech->version check done in scan_peb() since it's
already done when ubi_io_read_ec_hdr() is called.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>